First Phase Refactor: file system rearrange
