+++ CMakeLists.txt
@@ -1,0 +1,4 @@
+cmake_minimum_required(VERSION 3.18.4)
+
+add_subdirectory(GameSystem)
+#add_subdirectory(MapCreator)
