+++ GameSystem/img/compass.png
